keylogger_project_telegram_notification/requirements.txt
pynput
python-telegram-bot==13.15
schedule